Tests which require an IOC are renamed and incorporated in the testsuite
