+++ test/com/cosylab/epics/caj/test/AllTests.java
@@ -41,10 +41,10 @@ import gov.aps.jca.dbr.DBR_Enum;
     CAJContextRepeaterTest.class,
     CAJContextStateTest.class,
     CAJContextConfigurationTest.class,
-    CAJContextListenersTest.class,
+    CAJContextListenersIT.class,
     CAJContextChannelIT.class,
     CAJChannelIT.class,
-    CAJMonitorTest.class,
+    CAJMonitorIT.class,
     CAJContextPendPoolFlushIT.class,
     CAJChannelGetIT.class,
     CAJChannelPutIT.class,
